create a Common/Simulation to be used by Fatras and Geant4
