+++ Examples/Run/Fatras/Common/Fatras.cpp
@@ -22,6 +22,7 @@
 #include "ActsExamples/MagneticField/MagneticFieldOptions.hpp"
 #include "ActsExamples/Options/CommonOptions.hpp"
 #include "ActsExamples/Options/ParticleGunOptions.hpp"
+#include "ActsExamples/Simulation/CommonSimulation.hpp"
 #include "ActsExamples/TruthTracking/ParticleSelector.hpp"
 #include "ActsExamples/Utilities/Paths.hpp"
 
@@ -34,142 +35,19 @@ using namespace ActsExamples;
 
 namespace {
 
-// collection names
-static constexpr const char* kParticlesInput = "particles_input";
-static constexpr const char* kParticlesSelection = "particles_selection";
-static constexpr const char* kParticlesInitial = "particles_initial";
-static constexpr const char* kParticlesFinal = "particles_final";
-static constexpr const char* kSimHits = "simhits";
-
-// input handling
-
-void addInputOptions(ActsExamples::Options::Description& desc) {
-  using boost::program_options::value;
-
-  auto opt = desc.add_options();
-  // do not use the common input options: they have the wrong doc text and add a
-  // lot of unused options.
-  opt("input-dir", value<std::string>(),
-      "Read particle input from CSV files in the given directory. If not "
-      "given, particles are generated with the particle gun");
-  ActsExamples::Options::addParticleGunOptions(desc);
-  ActsExamples::ParticleSelector::addOptions(desc);
-}
-
-void setupInput(
-    const ActsExamples::Options::Variables& vars,
-    ActsExamples::Sequencer& sequencer,
-    std::shared_ptr<const ActsExamples::RandomNumbers> randomNumbers) {
-  auto logLevel = Options::readLogLevel(vars);
-
-  if (not vars["input-dir"].empty()) {
-    // read particle input from csv file
-    CsvParticleReader::Config readParticles;
-    readParticles.inputDir = vars["input-dir"].as<std::string>();
-    readParticles.inputStem = "particles";
-    readParticles.outputParticles = kParticlesInput;
-    sequencer.addReader(
-        std::make_shared<CsvParticleReader>(readParticles, logLevel));
-
-  } else {
-    // generate particle input from a particle gun
-    auto gen = Options::readParticleGunOptions(vars);
-    gen.outputParticles = kParticlesInput;
-    gen.randomNumbers = randomNumbers;
-    sequencer.addReader(std::make_shared<EventGenerator>(gen, logLevel));
-  }
-
-  // add additional particle selection
-  auto select = ActsExamples::ParticleSelector::readConfig(vars);
-  select.inputParticles = kParticlesInput;
-  select.outputParticles = kParticlesSelection;
-  sequencer.addAlgorithm(
-      std::make_shared<ActsExamples::ParticleSelector>(select, logLevel));
-}
-
-// output handling
-
-// output options are just the common output options
-
-void setupOutput(const ActsExamples::Options::Variables& vars,
-                 ActsExamples::Sequencer& sequencer) {
-  auto logLevel = Options::readLogLevel(vars);
-  auto outputDir =
-      ensureWritableDirectory(vars["output-dir"].template as<std::string>());
-
-  // Write simulation information as CSV files
-  if (vars["output-csv"].template as<bool>()) {
-    // NOTE: the collection names are an internal implementation detail but the
-    //   output file stems are a public user interface. the former can be
-    //   changed at-will, but the later should always be the same.
-
-    // write simulated particle initial states
-    CsvParticleWriter::Config writeInitial;
-    writeInitial.inputParticles = kParticlesInitial;
-    writeInitial.outputDir = outputDir;
-    writeInitial.outputStem = "particles_initial";
-    sequencer.addWriter(
-        std::make_shared<CsvParticleWriter>(writeInitial, logLevel));
-
-    // write simulated particle final states
-    CsvParticleWriter::Config writeFinal;
-    writeFinal.inputParticles = kParticlesFinal;
-    writeFinal.outputDir = outputDir;
-    writeFinal.outputStem = "particles_final";
-    sequencer.addWriter(
-        std::make_shared<CsvParticleWriter>(writeFinal, logLevel));
-
-    // write simulated hits
-    CsvSimHitWriter::Config writeSimHits;
-    writeSimHits.inputSimHits = kSimHits;
-    writeSimHits.outputDir = outputDir;
-    writeSimHits.outputStem = "hits";
-    sequencer.addWriter(
-        std::make_shared<CsvSimHitWriter>(writeSimHits, logLevel));
-  }
-
-  // Write simulation information as ROOT files
-  if (vars["output-root"].template as<bool>()) {
-    // NOTE: the collection names are an internal implementation detail but the
-    //   output file names are a public user interface. the former can be
-    //   changed at-will, but the later should always be the same.
-
-    // write simulated particle initial states
-    RootParticleWriter::Config writeInitial;
-    writeInitial.inputParticles = kParticlesInitial;
-    writeInitial.filePath = joinPaths(outputDir, "particles_initial.root");
-    sequencer.addWriter(
-        std::make_shared<RootParticleWriter>(writeInitial, logLevel));
-
-    // write simulated particle final states
-    RootParticleWriter::Config writeFinal;
-    writeFinal.inputParticles = kParticlesFinal;
-    writeFinal.filePath = joinPaths(outputDir, "particles_final.root");
-    sequencer.addWriter(
-        std::make_shared<RootParticleWriter>(writeFinal, logLevel));
-
-    // write simulated hits
-    RootSimHitWriter::Config writeHits;
-    writeHits.inputSimHits = kSimHits;
-    writeHits.filePath = joinPaths(outputDir, "hits.root");
-    sequencer.addWriter(
-        std::make_shared<RootSimHitWriter>(writeHits, logLevel));
-  }
-}
-
 // simulation handling
 
-void setupSimulation(
+void setupFatrasSimulation(
     const ActsExamples::Options::Variables& vars,
     ActsExamples::Sequencer& sequencer,
     std::shared_ptr<const ActsExamples::RandomNumbers> randomNumbers,
     std::shared_ptr<const Acts::TrackingGeometry> trackingGeometry) {
   auto logLevel = Options::readLogLevel(vars);
   auto fatrasCfg = FatrasSimulation::readConfig(vars);
-  fatrasCfg.inputParticles = kParticlesSelection;
-  fatrasCfg.outputParticlesInitial = kParticlesInitial;
-  fatrasCfg.outputParticlesFinal = kParticlesFinal;
-  fatrasCfg.outputSimHits = kSimHits;
+  fatrasCfg.inputParticles = Simulation::kParticlesSelection;
+  fatrasCfg.outputParticlesInitial = Simulation::kParticlesInitial;
+  fatrasCfg.outputParticlesFinal = Simulation::kParticlesFinal;
+  fatrasCfg.outputSimHits = Simulation::kSimHits;
   fatrasCfg.randomNumbers = randomNumbers;
   fatrasCfg.trackingGeometry = trackingGeometry;
   fatrasCfg.magneticField = ActsExamples::Options::readMagneticField(vars);
@@ -180,8 +58,10 @@ void setupSimulation(
 
 }  // namespace
 
-// fatras main function
-
+/// Fatras main function
+///
+/// Standard arguments @param argc and @param argv[] are forwarded
+/// @param detector abstracts the used detector input
 int runFatras(int argc, char* argv[],
               std::shared_ptr<ActsExamples::IBaseDetector> detector) {
   using namespace ActsExamples;
@@ -190,7 +70,7 @@ int runFatras(int argc, char* argv[],
   auto desc = Options::makeDefaultOptions();
   Options::addSequencerOptions(desc);
   Options::addRandomNumbersOptions(desc);
-  addInputOptions(desc);
+  Simulation::addInputOptions(desc);
   Options::addOutputOptions(desc, OutputFormat::Root | OutputFormat::Csv);
   // add general and detector-specific geometry options
   Options::addGeometryOptions(desc);
@@ -217,9 +97,9 @@ int runFatras(int argc, char* argv[],
     sequencer.addContextDecorator(cdr);
   }
   // setup algorithm chain
-  setupInput(vars, sequencer, randomNumbers);
-  setupSimulation(vars, sequencer, randomNumbers, trackingGeometry);
-  setupOutput(vars, sequencer);
+  Simulation::setupInput(vars, sequencer, randomNumbers);
+  setupFatrasSimulation(vars, sequencer, randomNumbers, trackingGeometry);
+  Simulation::setupOutput(vars, sequencer);
 
   // run the simulation
   return sequencer.run();
